requirements.txt
Flask==2.3.2
Werkzeug>=2.0,<3.0
Flask-Cors==3.0.10
SQLAlchemy==1.4.47
requests==2.28.2
beautifulsoup4==4.12.2
openai==0.27.8
python-dotenv==1.0.0
pytest==7.2.0
pytest-flask==1.2.0
alembic==1.11.1